Remove outdated code

+++ reckoning.py
@@ -18,30 +18,39 @@ print("Last csv:")
 print(latestCSV)
 print("Due csvs:")
 print(dueFiles)
+dueAnswer = input("Is this correct? [Y/N] ")
 
-yesterdayDate = todayDate.shift(days=-1)
-filenameYesterday = f"ES_{yesterdayDate.format('DD')}-{yesterdayDate.format('MM')}.csv"
-filenameToday = f"ES_{todayDate.format('DD')}-{todayDate.format('MM')}.csv"
-
-shutil.move(latestCSV, filenameYesterday)
-
-for file in dueFiles:
-    # Copy oldest MICRODADOS to download folder (overwrite)
-    currentCSV = dueFiles.pop(0)
-    print("Arquivo:", currentCSV)
-    shutil.copy(currentCSV, Path("/home/atilioa/Downloads/MICRODADOS.csv"))
-    # Run posta_relatorio.py with -t
-    subprocess.run(["python", "posta_relatorio.py", "-t"])
-    # Rename desktop result file to actual day
-    newCSVFilename = f'ES_{currentCSV[-6:-4]}-{yesterdayDate.format("MM")}.csv'
-    print(currentCSV, newCSVFilename)
-    os.rename(
-        Path(os.path.join("/home/atilioa/Área de Trabalho", filenameToday)),
-        Path(os.path.join("/home/atilioa/Área de Trabalho", newCSVFilename)),
+if dueAnswer.lower() in ["y", "s", "sim", "yes"]:
+    yesterdayDate = todayDate.shift(days=-1)
+    filenameYesterday = (
+        f"ES_{yesterdayDate.format('DD')}-{yesterdayDate.format('MM')}.csv"
     )
-    try:
-        os.remove(Path(os.path.join("/home/atilioa/Área de Trabalho", currentCSV)))
-    except:
-        pass
-    # Rename current directory result file to yesterday (overwrite)
-    shutil.copy(filenameToday, filenameYesterday)
+    filenameToday = f"ES_{todayDate.format('DD')}-{todayDate.format('MM')}.csv"
+
+    shutil.move(latestCSV, filenameYesterday)
+
+    for file in dueFiles:
+        # Copy oldest MICRODADOS to download folder (overwrite)
+        currentCSV = dueFiles.pop(0)
+        print("Arquivo:", currentCSV)
+        shutil.copy(currentCSV, Path("/home/atilioa/Downloads/MICRODADOS.csv"))
+        
+        # Run posta_relatorio.py
+        subprocess.run(["python", "posta_relatorio.py", "-d", "f"])
+        
+        # Rename desktop result file to actual day
+        newCSVFilename = f'ES_{currentCSV[-6:-4]}-{yesterdayDate.format("MM")}.csv'
+        print(currentCSV, newCSVFilename)
+        
+        os.rename(
+            Path(os.path.join("/home/atilioa/Área de Trabalho", filenameToday)),
+            Path(os.path.join("/home/atilioa/Área de Trabalho", newCSVFilename)),
+        )
+        
+        try:
+            os.remove(Path(os.path.join("/home/atilioa/Área de Trabalho", currentCSV)))
+        except:
+            pass
+        
+        # Rename current directory result file to yesterday (overwrite)
+        shutil.copy(filenameToday, filenameYesterday)
